Implementação de testes unitários para login e logout da aplicação.
